Fixed issues where Binary could change and cause problems
Fixed issue with binary being set to 0bytes through pipe error.

Fixed issue where if the user deleted the binary from the system it would not re-check for it on startup and in app.

Now the binary is checked every time I create the child process.

Internet must be enabled if the binary is not present currently.